refactor(tile): move the attention verbs out of mod.rs
`mod.rs` declares modules and re-exports; the code sits in `base.rs`, as every
other module here does. Pure move.
